propback/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
Django==4.1
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-filter==22.1
django-multiselectfield==0.1.12
django-taggit==3.0.0
djangorestframework==3.13.1
idna==3.3
ndg-httpsclient==0.5.1
Pillow==9.2.0
pusher==3.2.0
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==22.0.0
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12